captcha-solver/requirements.txt
pytesseract==0.3.7
Pillow==8.2.0
Flask==2.0.1
flask-cors==3.0.10
tornado==6.1
